Add netbird-bin package
Only contains the binary release for amd64 atm. Have not had time to
figure out the build process / add more arches.

Will also probably be a little buggy still. There's a system service
included which needs to run for netbird to work.

Setup is a tiny bit cumbersome in that you have to enable the service
_first_ (which will complain about no auth found) and then manually run
`netbird up` with your pre-shared key or sso or whatever method you
choose to authenticate.